requirements/base.txt
pip # get rid of annoying pip messages

# Django
# ------------------------------------------------------------------------------
django>=2.2,<2.3 # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
argon2-cffi==18.1.0  # https://github.com/hynek/argon2_cffi
django-solo
slackclient
psycopg2-binary